fix: modal 폴더 내 컴포넌트의 스타일 파일 분리
